Step 4/visual-pdf-classifier-main/requirements.txt
fitz==0.0.1.dev2
imagesize==1.3.0
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.2
pdf2image==1.16.0
Pillow==9.0.1
psycopg2==2.9.3
pytorch_lightning==1.6.0
PyYAML==6.0
scikit_learn==1.0.2
timm==0.5.4
torch==1.11.0
torchmetrics==0.7.3
torchvision==0.12.0
tqdm==4.62.3
